+++ lib/nokaya/app.rb
@@ -48,6 +48,16 @@ module Nokaya
       download_and_save img_link, nokaya
     end
 
+    desc "imgur_album", "Get all images from an Imgur album (nokaya -ial url)"
+    map "-ial" => :imgur_album
+    def imgur_album *args
+      abort Status.no_url if args.empty?
+      nokaya = Getter.new options, :imgur_album, args
+      page = nokaya.parse_page
+      img_links = nokaya.get_imgur_album page
+      download_album img_links, nokaya
+    end
+
     private
 
     def download_and_save img_link, nokaya
@@ -55,6 +65,17 @@ module Nokaya
       path = Image.photo_name nokaya
       Image.save_image(path, nokaya.get_image(img_link))
       puts Status.saved path
+    end
+
+    def download_album img_links, nokaya
+      puts "\nDownloading album at #{nokaya.url}...\n\n"
+      img_links.each do |link|
+        parsed = URI.parse link
+        file = "#{Dir.home}/Downloads#{parsed.path}"
+        puts "Saving #{file}\n"
+        Image.save_image(file, nokaya.get_image(link))
+      end
+      puts "\nDone.\n\n"
     end
 
   end